requirements.txt
streamlit==1.37.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==10.4.0
scikit-image==0.24.0
